Handle image commands like glyph commands
* Add lilyglyphs:add_image_command() and lilyglyphs:add_image_commands()
* Factor out common code with glyph commands
* Provide \lilyglyphs (logo) as first test command.

NOTE: This is only a first step that will at least allow me
to reimplement the commands from the previous implementation.
For later there has to be support for storing the LilyPond source
in the Formatter Entry Table and to rebuild the image if it is
not found (have to provide a path in the personal texmf tree).